+++ mbus/mbus-protocol-aux.h
@@ -275,7 +275,7 @@ int mbus_send_application_reset_frame(mbus_handle * handle, int address, int sub
  * 
  * @return Zero when successful.
  */
-int mbus_send_switch_baudrate_frame(mbus_handle * handle, int address, int baudrate);
+int mbus_send_switch_baudrate_frame(mbus_handle * handle, int address, long baudrate);
 
 /** 
  * Sends request frame (REQ_UD2) to given slave using "unified" handle
